Update the tutorial to follow the latest version
The problem which have been corrected here are:

1) Many function which required or returned `String` require or return
`Text` now. I fixed these instances both in the code and in the text.

2) In the last version there was a deprecation of some of the monadic
functions in the interface, like `mapDyn` or `combineDyn` in favour of
`fmap` and `zipDynWith`. Again, this has been fixed both in the code and
in the text of the tutorial.

I actually typechecked all the new fragments to make sure they
typecheck, but I might have missed something, so please double-check.