Companion PR for reflex-dom PR: Ask user for permissions.
From Android 6 onwards we have to ask the user for permissions at runtime. This
patch provides the means for the WebView handling code to ask the user for
permissions for accessing the camera or the microphone.

This patch is necessary, because Java can't simply have callbacks. Instead
Activity has a method onRequestPermissionsResult() which has to be overridden to
get the result, of asking the user, back.