Cleanup splicing logic
This moves the "spliced" ghc packages to its own package set. This
will make things easier so as not to get confused between vanilla GHC
HEAD and the patched version.

In addition the package set used was not exactly correct previously so
the build packages are hopefully set up better.

Reworked haskell-overlays logic to use ghcSavedSplices.

Added a SPLICE_DIR attribute to more accurately get the directory to
use in -save-splices & -load-splices.