Disable correct optimizations in text-jsstring
I had to turn on these flags to get text-jsstring to work.
Unfortunately some of them may be useful - but they interfere with
jsstring inlining. Hopefully in the future we can figure out why they
break inlining & fix the issue either in GHC or GHCJS.

- specialise
- full-laziness
- float-in

In addition, setting -fignore-interface-pragmas seems to be necessary.