Update ghcjs to -no-string-compact
This adds a flag to disable the string compacting done in ghcjs 8.4.
This includes the "string packing" done. We want to avoid this in
text-jsstring reflex-platform for performance reasons.

These commits should go into obsidiansystems/ghcjs but I don’t have
write access. Hopefully can fix that tomorrow.

Notes on disabling string compacting... It doesn’t seem to break
anything currently but it looks like we do get duplicate strings in
some places. The generated code is larger but not ridiculously and I
think that the minifier + closure compiler can fix most of the bloat.