Fix text override for ghcjs-8.4
`text` is now a boot package since ghc 8.4. This means that `useTextJSString`
which overrides `text` to depend on `ghcjs-base` and use `JSString` under the
hood now needs to happen before ghcjs booting when enabled until there is
better story for boot pkgs in nixpkgs or text-jsstring patches are done
differently. This now adds `ghcjs-base` and it's deps dlist, vector and
primitive as part of the boot packages. All of this only has effect on ghcjs
when `useTextJSString` is enabled.