th-splicing: Handle patching abi hashes of different lengths
reflex(possibly others?) has a hash of length 20 when building splices vs a length
of 21 when building on cross which requires also updating the prefix bytes that
include the length when patching the splices. Ideally there would be a more robust
way to update the abi hashes used in th splices.