Stop nulling out some packages for GHCJS
This was done because they kept on trying to be built for Setup.hs. That
problem is fixed properly now, so it should be safe to remove these.

Additionally, hscolour is also a library that could be used by a GHCJS
app. By making this libary availible, this cleanup provides a practical
benefit. Semantic-reflex-example in particular uses hscolour as a
library.