Prioritize *.cabal over default.nix for work-on
The work-on script uses package_invocation to find a package. This
will default to finding the default.nix script. Some users like in
 #101 don’t want work-on to pick up the default.nix script as they are
using default.nix for other reasons. To fix this, we can just
prioritize Cabal over default.nix. The scripts will still look in
default.nix if there are no cabal files available. Because cabal2nix
has gotten pretty good, it is a pretty safe idea in this case.

Fixes #101