Move hackGet to lib directory
These aren’t overlays, they should go in the lib/ directory. This
eliminates the dependency of reflex-platform/haskell-overlays on the
unexposed hack-get overlay. Instead, we provide it as a library
function. “lib” is intended to be a place to put other reflex-platform
helper functions that we use. These should only depend on Nixpkgs.