Only include /bin of build tools
Some of our build tools like ghcid are themselves Haskell packages. As
a result, their propagated dependencies would wind up in the
package.conf.d when strictDeps = false. This leads to an inconsistency
between “ob run” which uses these build tools and everything else. To
avoid exposing the entire build tools package, we wrap each so that
only /bin/* is available in the shell.