Don't rebuild haskell-ide-engine to project overrides
In 4f0d01963de36f2691d0ee03e892993c0473f721 we made
`nativeHaskellPackages` always close over the `ghc` that came with
reflex-platform. This means that app-specific overrides won't cause the
needless rebuilding of these tools. `haskell-ide-engine` however wasn't
using this, and so would be rebuilt.

This fixes everything to use `ghc` by making `nativeHaskellPackages` a
pattern argument so it can have `ghc` as a default.