Support both Apple Development and iPhone Developer certificates.
The previous fix was just ignoring the first `-c`, this contatenates the outputs to
find either versions. On single developer machines, this will work fine as there is only one
but on shared machines where there are keys in the system keychain and not only in the login
keychain users will be unable to deploy until they delete all others from the system keychain
and that has to be done by the user that put it in the first place.

A fix for multi user systems would be to accept the developer id in either default.nix or
to accept it in the deploy script.