Verify .json is only file in directory for hackGet
This throws an error when the git/github.json and default.nix file are
not the only file(s) in the thunk directory. This is a common issue
when users accidentally forget to remove the .json after they’ve
unpacked a thunk. Note that default.nix is optional and occasionally
provided in repo’s that are hackGet’d.