Disable multi-dex on Android builds
Multi-dex allows your app to have more than 64k methods but
reflex-platform apps tend to have very few methods since most logic is
not in Java. Multi-dex also causes problems when classes are not
obviously dependencies of the main class which is common in
reflex-platform apps because it uses JNI. The complexity of teaching
the build how to keep the right classes in the main dex file is not
worth the benefit.

See https://developer.android.com/studio/build/multidex.html for
details.