enableLibraryProfiling shouldn't change the compiler
The enableLibraryProfiling flag has historically only changed how we
compile libraries that are not bundled with ghc. There's a new flag
("enableProfiledLibs") in nixpkgs for ghc 8.10+ that allows one to set
to disable production of the profiled versions of bundled ghc libraries.
By always setting enableProfiledLibs to true, we retain the behavior we
had with ghc 8.6.5 and the default behavior of binary distributions of
ghc.