ghc 8.10 `override` then `overrideAttrs`
I see there is some tricks in Nixpkgs so the order doesn't actually
matter, but I rather not rely on them.

`override` changes the arguments to a `callPackaged` function.
`overrideAttrs` changes the result *of that function application*.
`override` first so we know what value we are modifying just makes mose
sense.