aarch64-linux: avoid tools that make the build fail
some mix of these tools are pulling in ghc8.6 compiler which actually
ends up failing to build. It may be due to amount of ram in the device I
have, but didn't follow too closely.

These tools are optional and nice-to-have so for now I will just be
removing them