Revert "reflex-dom-core: disable tests to fix determinism"
We should look at the actual error and fix it. There are some hard
coded wait times in the tests, if they're causing the failure we can
increase them somewhat. AFAICT candidates functions to adjust are
`withSeleniumServer` and `withRetry`.

Running the `gc` test with `cpulimit -l 1` results in:
```
...
438664
436864
440056
SUCCEEDED
[0207/130124.616315:ERROR:broker_posix.cc(40)] Recvmsg error:
Connection reset by peer (104)
```
which may also be considered a failure by the test suite.

This reverts commit f9a54be4fe65f1fed1e22d0abd186e63233fb21a.