Bump `dependent-sum` and `dependent-sum-template`
We were previously missing the latest version of
`dependent-sum-template` due to confusion over them being in the same
repo. Now, they are split into separate repos again, something which
reduces cognitive overhead *and* is now safe because neither dependson
the other. (They both instead depend on `some`.)

We want to release both, so making this PR to CI everything with the
release candidates. If this passes, we will cut release and them bump
the thunks again to point to `master` not `develop`.