Remove deprecated android settings
Fixes warnings:

WARNING:: The option setting 'android.enableBuildCache=false' is deprecated.
The current default is 'true'.
It will be removed in version 7.0 of the Android Gradle plugin.
It does not do anything and AGP is now using Gradle caching.

WARNING:: The option setting 'android.enableR8=false' is deprecated.
It will be removed in version 7.0 of the Android Gradle plugin.
You will no longer be able to disable R8